Change user id to Long

+++ server/src/main/java/ovh/nixenos/tab/server/dto/user/UserDTOInput.java
@@ -1,10 +1,8 @@
 package ovh.nixenos.tab.server.dto.user;
 
-import ovh.nixenos.tab.server.users.User;
-
 public class UserDTOInput {
 
-  private Integer id;
+  private Long id;
 
   /*
    * username of the user
@@ -44,11 +42,11 @@ public class UserDTOInput {
   public UserDTOInput() {
   }
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
